+++ shardis-ui/config/spec-bundle.js
@@ -9,7 +9,6 @@
 */
 Error.stackTraceLimit = Infinity;
 require('core-js');
-require('ts-helpers');
 
 require('zone.js/dist/zone');
 require('zone.js/dist/long-stack-trace-zone');