+++ shardis-ui/src/main/frontend/polyfills.ts
@@ -1,7 +1,6 @@
 // Ccre.js
 import 'core-js/es6';
 import 'core-js/es7/reflect';
-import 'ts-helpers';
 
 // Zone.js
 require('zone.js/dist/zone');